combine all vis under single test
