+++ superset/assets/cypress/integration/explore/visualizations/_box_plot.js
@@ -1,18 +1,17 @@
-describe('Treemap', () => {
-  const TREEMAP_FORM_DATA = {
+export default () => describe('Box Plot', () => {
+  const BOX_PLOT_FORM_DATA = {
     datasource: '2__table',
-    viz_type: 'treemap',
-    slice_id: 50,
+    viz_type: 'box_plot',
+    slice_id: 49,
     granularity_sqla: 'year',
     time_grain_sqla: 'P1D',
     time_range: '1960-01-01+:+now',
     metrics: ['sum__SP_POP_TOTL'],
     adhoc_filters: [],
-    groupby: ['country_code'],
-    row_limit: 50000,
+    groupby: ['region'],
+    limit: '25',
     color_scheme: 'bnbColors',
-    treemap_ratio: 1.618033988749895,
-    number_format: '.3s',
+    whisker_options: 'Min/max+(no+outliers)',
   };
 
   function verify(formData) {
@@ -27,22 +26,13 @@ describe('Treemap', () => {
   });
 
   it('should work', () => {
-    verify(TREEMAP_FORM_DATA);
-    cy.get('.chart-container svg rect.child').should('have.length', 214);
-  });
-
-  it('should work with multiple groupby', () => {
-    verify({
-      ...TREEMAP_FORM_DATA,
-      groupby: ['region', 'country_code'],
-    });
-    cy.get('.chart-container svg rect.parent').should('have.length', 7);
-    cy.get('.chart-container svg rect.child').should('have.length', 214);
+    verify(BOX_PLOT_FORM_DATA);
+    cy.get('.chart-container svg rect.nv-boxplot-box').should('have.length', 7);
   });
 
   it('should work with filter', () => {
     verify({
-      ...TREEMAP_FORM_DATA,
+      ...BOX_PLOT_FORM_DATA,
       adhoc_filters: [{
         expressionType: 'SIMPLE',
         subject: 'region',
@@ -54,6 +44,7 @@ describe('Treemap', () => {
         filterOptionName: 'filter_8aqxcf5co1a_x7lm2d1fq0l',
       }],
     });
-    cy.get('.chart-container svg rect.child').should('have.length', 8);
+    cy.get('.chart-container svg rect.nv-boxplot-box').should('have.length', 1);
   });
+
 });
