+++ superset/assets/cypress/integration/explore/visualizations/_treemap.js
@@ -1,17 +1,18 @@
-describe('Box Plot', () => {
-  const BOX_PLOT_FORM_DATA = {
+export default () => describe('Treemap', () => {
+  const TREEMAP_FORM_DATA = {
     datasource: '2__table',
-    viz_type: 'box_plot',
-    slice_id: 49,
+    viz_type: 'treemap',
+    slice_id: 50,
     granularity_sqla: 'year',
     time_grain_sqla: 'P1D',
     time_range: '1960-01-01+:+now',
     metrics: ['sum__SP_POP_TOTL'],
     adhoc_filters: [],
-    groupby: ['region'],
-    limit: '25',
+    groupby: ['country_code'],
+    row_limit: 50000,
     color_scheme: 'bnbColors',
-    whisker_options: 'Min/max+(no+outliers)',
+    treemap_ratio: 1.618033988749895,
+    number_format: '.3s',
   };
 
   function verify(formData) {
@@ -26,13 +27,22 @@ describe('Box Plot', () => {
   });
 
   it('should work', () => {
-    verify(BOX_PLOT_FORM_DATA);
-    cy.get('.chart-container svg rect.nv-boxplot-box').should('have.length', 7);
+    verify(TREEMAP_FORM_DATA);
+    cy.get('.chart-container svg rect.child').should('have.length', 214);
+  });
+
+  it('should work with multiple groupby', () => {
+    verify({
+      ...TREEMAP_FORM_DATA,
+      groupby: ['region', 'country_code'],
+    });
+    cy.get('.chart-container svg rect.parent').should('have.length', 7);
+    cy.get('.chart-container svg rect.child').should('have.length', 214);
   });
 
   it('should work with filter', () => {
     verify({
-      ...BOX_PLOT_FORM_DATA,
+      ...TREEMAP_FORM_DATA,
       adhoc_filters: [{
         expressionType: 'SIMPLE',
         subject: 'region',
@@ -44,7 +54,6 @@ describe('Box Plot', () => {
         filterOptionName: 'filter_8aqxcf5co1a_x7lm2d1fq0l',
       }],
     });
-    cy.get('.chart-container svg rect.nv-boxplot-box').should('have.length', 1);
+    cy.get('.chart-container svg rect.child').should('have.length', 8);
   });
-
 });
