+++ superset/assets/cypress/integration/explore/visualizations/all.js
@@ -1,15 +1,35 @@
+import AreaTest from './_area';
+import BigNumberTest from './_big_number';
 import BigNumberTotalTest from './_big_number_total';
+import BubbleTest from './_bubble';
 import CompareTest from './_compare';
 import DistBarTest from './_dist_bar';
+import DualLineTest from './_dual_line';
 import HistogramTest from './_histogram';
 import LineTest from './_line';
+import PieTest from './_pie';
+import PivotTableTest from './_pivot_table';
+import SankeyTest from './_sankey';
+import SunburstTest from './_sunburst';
 import TableTest from './_table';
+import TreemapTest from './_treemap';
+import WorldMapTest from './_world_map';
 
 describe('All Visualizations', () => {
+  AreaTest();
+  BigNumberTest();
   BigNumberTotalTest();
+  BubbleTest();
   CompareTest();
   DistBarTest();
+  DualLineTest();
   HistogramTest();
   LineTest();
+  PieTest();
+  PivotTableTest();
+  SankeyTest();
+  SunburstTest();
   TableTest();
+  TreemapTest();
+  WorldMapTest();
 });
